+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## [Unreleased]
 ### Added
 - Inherited factory methods are now found and used (for Java-based tests, inheritance of static methods not supported in Kotlin) 
-- Compatibility tests for build systems (Maven, Gradle) and application frameworks (Quarkus, SpringBoot)
+- Compatibility tests for build systems (Maven, Gradle) and frameworks (JUnit, Quarkus, SpringBoot)
 ### Changed
 - Easier install as dependent JUnit modules are no longer packaged with TableTest distribution
 - TableTestException now of type ParameterResolutionException